django_core/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
django-cors-headers
celery
redis
boto3
pyjwt
kafka-python
python-dotenv
cryptography
pydantic